Fix Flatpak build jobs
[skip ci]

Flatpak increased the metadata size limit to 128 MB in:
https://github.com/ostreedev/ostree/commit/4bac96a8c817beda37448f9b8c662162bb619981

However, that upstream change is not available in Ubuntu 18.04 or 20.04.

Since rustdesk-org does not provide a Dockerfile for x86_64 based on Ubuntu 22.04,
this commit switches the source of `run-on-arch-action` to our own repository,
which includes the necessary Dockerfile.

The affected jobs now use Ubuntu 22.04, allowing them to complete successfully.

Signed-off-by: Manuel Fombuena <mfombuena@innovara.tech>